Discordbot/requirements.txt
discord.py
requests
python-decouple
datetime